[Feature] Adding gauge to represent fuzzy indicator
