+++ package.json
@@ -55,6 +55,7 @@
     "react-datepicker": "^3.2.2",
     "react-dev-utils": "^10.2.1",
     "react-dom": "^16.13.1",
+    "react-gauge-chart": "^0.2.5",
     "react-icons": "^3.11.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
